+++ controllers/ApiController.php
@@ -23,6 +23,9 @@ use yii\helpers\Url;
 use humhub\modules\user\models\User;
 use humhub\modules\onlyoffice\notifications\Mention as Notify;
 use humhub\modules\onlyoffice\models\Mention;
+use humhub\modules\content\models\ContentContainer;
+use humhub\modules\user\models\User;
+use humhub\modules\cfiles\permissions\ManageFiles;
 
 class ApiController extends Controller
 {
@@ -140,6 +143,59 @@ class ApiController extends Controller
 
         return $this->asJson([
             'file' => $file
+        ]);
+    }
+
+    /**
+     * Rename action
+     */
+    public function actionRename()
+    {
+        if (($body_data = file_get_contents('php://input')) === FALSE) {
+            throw new \Exception('Empty body');
+        }
+
+        $data = json_decode($body_data, TRUE);
+        if ($data === NULL) {
+            throw new \Exception('Could not parse json');
+        }
+
+        $file = File::findOne(['onlyoffice_key' => $data['key']]);
+        
+        $owner = User::findOne($file->created_by);
+        $containerRecord = ContentContainer::findOne(['id' => $owner->contentcontainer_id]);
+        $container = $containerRecord->getPolymorphicRelation();
+        $canRename = $container->can(ManageFiles::class);
+        if(!$canRename) {
+            throw new \Exception('Permission denied');
+        }
+
+        $newFileName = $data['newFileName'];
+        $origExt = $data['ext'];
+        $arrayName = explode(".", $newFileName);
+        $curExt = end($arrayName);
+
+        if($origExt !== $curExt) {
+            $newFileName .= "." . $origExt;
+        }
+
+        $file->updateAttributes(['file_name' => $newFileName]);
+
+        $meta = [
+            "c" => "meta",
+            "key" => $data['key'],
+            "meta"=> [
+                "title" => $newFileName
+            ]
+        ];
+        $response = $this->module->commandService($meta);
+
+        if($response['error'] !== 0){
+            throw new \Exception('Error from command Service: ' . $response['error']);
+        }
+
+        return $this->asJson([
+            'file' => FileHelper::getFileInfos($file)
         ]);
     }
 }