+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 - editor customization
 - validation of server settings on the settings page
+- renaming from editor
 - keep intermediate versions when editing (forcesave)
 - connect to the demo server
 - trial period of 30 days for the demo server